distributed-task-manager/services/tasks-service/requirements.txt
flask
flask-pymongo
PyJWT
bson
